pyproject.toml
[project]
name = "stock-bot"
version = "0.2.0"
description = "Telegram bot for marciano inventory management via Google Sheets"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.49.1",
    "gspread>=6.2.1",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
]
